requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-common==1.1.28
azure-core==1.25.1
azure-functions==1.11.2
azure-identity==1.11.0
azure-keyvault-secrets==4.6.0
azure-storage-blob==12.13.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
DateTime==4.7
html5lib==1.1
idna==3.4
isodate==0.6.1
lxml==4.9.1
msal==1.19.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.21.6
oauthlib==3.2.1
pandas==1.3.5
Pillow==9.2.0
portalocker==2.5.1
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.4
pywin32==304
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
types-cryptography==3.3.23
typing-extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
XlsxWriter==3.0.3
zope.interface==5.4.0
